renamed App.js to App.jsx and corrected some styles issues
