+++ src/index.js
@@ -3,6 +3,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import 'font-awesome/css/font-awesome.min.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import App from './App';
+import App from './App.jsx';
 
 ReactDOM.render(<App />, document.getElementById('root'));
